requirements.txt
Pillow==7.0.0
django==2.2.9
django-ckeditor==5.8.0
